+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Feature: Distinct for DSL Entity Select
 - Feature: Distinct for DSL SubQuery
 - Feature: Simple EnumConverter
+- Feature: InlineValue with Alias
 - Removal: Removed notLike and like
 - Refactor: Refactor of IT Tests
 - Bugfix: Fix missing upsert for entities with relationships